main schrijft nu goede csv

+++ Functies/calculate_score.py
@@ -5,4 +5,4 @@ def calculate_score(connection_nr, traject_nr, total_connections, total_time):
     p = connection_nr / total_connections
     T = traject_nr
     Min = total_time
-    return int(p * 10000 - (T * 100 + Min))
+    return float(p * 10000 - (T * 100 + Min))
